feat: engine-neutral xql.arrow_dataset() — Polars works for free
Move XarrayPushdownDataset and XarrayArrowStream to
backends/pyarrow.py: the pushdown dataset is a real
pyarrow.dataset.Dataset, so it serves every consumer of that protocol,
not just DuckDB. The new public constructor xql.arrow_dataset(ds)
makes that explicit:

    pl.scan_pyarrow_dataset(xql.arrow_dataset(ds))   # Polars, lazy
    con.register("t", xql.arrow_dataset(ds))         # DuckDB
    xql.arrow_dataset(ds).to_table(columns=..., filter=...)  # pyarrow

Verified with Polars 1.42: scans are lazy, predicates and projections
push into the dataset (a filtered group-by read 1 of 20 chunks and 3
of 5 columns), results match xarray exactly, and Polars frames
round-trip through xql.to_dataset via the Arrow PyCapsule protocol.

The DuckDB adapter is now a thin registration shim over the shared
module. polars added to the test extra.
