+++ xarray_sql/__init__.py
@@ -1,5 +1,5 @@
 from . import cftime
-from .backends import register
+from .backends import arrow_dataset, register
 from .df import from_map
 from .reader import read_xarray, read_xarray_table
 from .roundtrip import to_dataset
@@ -10,6 +10,7 @@ __all__ = [
     "XarrayContext",
     "read_xarray_table",
     "read_xarray",
+    "arrow_dataset",
     "register",
     "to_dataset",
     "from_map",  # deprecated
